+++ apps/presentations/state/history-controller.js
@@ -83,7 +83,7 @@
   }
 
   global.InkDeskPresentationsHistory = Object.freeze({
-    version: '0.20.2.9',
+    version: '0.20.2.10',
     create(options) {
       return new PresentationHistoryController(options);
     },